tests/wpt/web-platform-tests/css/css-forced-color-adjust/parsing/forced-color-adjust-invalid.html

<!DOCTYPE html>
<title>CSS Forced Color Adjust: parsing forced-color-adjust with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-color-adjust-1/#forced-color-adjust-prop">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test_invalid_value("forced-color-adjust", "auto auto");
  test_invalid_value("forced-color-adjust", "auto none");
  test_invalid_value("forced-color-adjust", "none none");
  test_invalid_value("forced-color-adjust", "none auto");
  test_invalid_value("forced-color-adjust", "1");
  test_invalid_value("forced-color-adjust", "revert");
  test_invalid_value("forced-color-adjust", "default");
</script>
